final_diploma/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-environ==0.7.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
flower==1.0.0
humanize==3.11.0
idna==3.2
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
model-bakery==1.3.2
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
ujson==4.2.0
uritemplate==3.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
